Exercise 4.19 added solution [Require login for POST]
