+++ chapter4/bloglist/models/user.js
@@ -14,7 +14,7 @@ const userSchema = new mongoose.Schema({
 userSchema.set("toJSON",  {
   transform: (document, returnedObject) => {
     returnedObject.id = returnedObject._id.toString()
-    // delete returnedObject.passwordHash
+    delete returnedObject.passwordHash
     delete returnedObject._id
     delete returnedObject.__v
     // delete returnedObject.blogpost // It may be usefull to comment, if blogs will be needed
